Abstract out sending usage updates, rename nc package to natsconn
